pyproject.toml
[project]
name = "ingram-checker"
version = "0.1.0"
description = "Ingram Lightning Source PDF compliance checker and fixer"
requires-python = ">=3.11"
dependencies = [
    "pikepdf>=9.0.0",
    "pymupdf>=1.24.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
]

[project.scripts]
ingram-check = "ingram_checker.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingram_checker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
